requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
monai~=1.3.1
numpy~=1.24.3
pytorch-lightning~=1.6.0
nibabel~=5.3.0